Move terminal-only worktree state into a per-kind TerminalSurfaceStore
WorktreeTerminalState's generic core (tree / tab / focus / occlusion /
zoom / persistence scheduling) no longer declares terminal-kind state.
The Ghostty view map, zmx launch/session bookkeeping, blocking-script
tracking, setup-script flag, environment injection, and agent-OSC
notification dedupe now live in TerminalSurfaceStore, together with the
launch-resolution and environment helpers that only touch that state.
Private accessors keep the orchestration bodies unchanged, so the diff
is the ownership move, not a rewrite. Per-kind stores are keyed by
surface/tab ID and empty when the kind is absent, so additional kinds
don't multiply per-worktree state.
